DAK-1216: Improve QUESO Cmake build
Knock out a few TODOs including:

 - expand determination of #defins using cmake probing
 - add support for a build version based on either a VERSION file
   generated as an additional step in the rsync process (see below)
   if present or on the current sha1 of the Dakota packages/external
   submodule if the VERSION file is not present
 - add content for CXXFLAGS output to queso.h

For now, the #defines do not assign a value except in a few places.
the Queso source code does not appear to check for a paraticlar value
(0 or 1) but rather only if the macro is defined.  This should work for
platforms except MSVC++ (which I don't think includes Queso).

The steps used to bring in a new Queso version into Dakota are now augmented
to include cration of a VERSION file as follows:

    git clone git@github.com:libqueso/queso.git queso-git
    cd queso-git
    git checkout origin/v${qversion}-release
    git branch -v |grep '^*' |awk '{print $2 " (" $3 ")"};' > VERSION

    rsync --recursive --delete --exclude=".git" --exclude="CMakeLists.txt" --exclude="config_queso.h.cmake.in" AUTHORS CHANGES COPYING LICENSE VERSION QUESO_users_manual.pdf README.md src test ../dakota-packages/queso/

There are a few remaining TODOs which require either interaction with
the maintainers of the native Queso repository and/or a newer minimum
version of cmake used in Dakota.