GPMSA: Workaround to send bounded domain to QUESO
This patch should not be maintained once we pull dev QUESO newer than
0.57.x (specifically commits from
https://github.com/libqueso/queso/pull/577)

The calibration parameter domain was getting hard wired to [0, 1]
causing a problem when the user's problem was on a different domain.
Since we know the prior's imageSet is a BoxSubset, we cast it and use
its bounds.