DAK-1464:  Made Acro/COBYLA symbols unique.
For the purposes of resolving symbol clashes with the NOWPAC/NLOPT
version of COBYLA, prepended all ACRO/COBYLA function/subroutines (and
calls to them) with "acro_".  Did this on devel, so it can be merged
into whatever branches are appropriate.  Tests are clean.  Also tested
the changes on the multilevel_harden branch.  NOWPAC/SNOWPAC now runs
without issues without having to disable ACRO in the build.