DP-370: Update pybind11 to v2.6.0
Address critical bug with pybind11 and Python 3.9.0 that is worked
around in pybind11 2.6.0 and will be fixed in Python 3.9.1.

From https://pybind11.readthedocs.io/en/stable/index.html:

Combining older versions of pybind11 (< 2.6.0) with the brand-new
Python 3.9.0 will trigger undefined behavior that typically manifests
as crashes during interpreter shutdown (but could also destroy your
data. You have been warned.)

We recommend that you wait for Python 3.9.1 slated for release in
December, which will include a fix that resolves this problem. In the
meantime, please update to the latest version of pybind11 (2.6.0 or
newer), which includes a temporary workaround specifically when Python
3.9.0 is detected at runtime.