Patch indexing in OPT++ native finite differencing (constraint FD routines for forward/backward/central differences).
This issue likely dates back to the conversion from Newmat to Teuchos, and
has been exposed in ACV numerical solves for linear objective functions +
nonlinear constraints.  Even when memory bounds were not exceeded, FD data
was being put in the wrong place.

Not advancing the Dakota sha1 yet, but this fixes the OPT++ solver problem
for the cherry-pick on online_debug_output.